irc: improve client disconnection
We now use io_uring to close the client, which removes the need to call
posix.shutdown. We do this by cancelling the read completion. Once
cancelled, we pick up an error in the onRead callback which closes the
connection and cleans it all up. This *should* get us to no syscalls in
the main thread after we have inited io_uring