auth: add atproto authorization
We take an app password and handle over SASL plain. We resolve the
handle back to the PDS server and perform auth there (including did:web
support). We use the createSession api, but store the hashed password
and the refresh token. As long as the password hash stays the same, we
just use the refresh token to authorize the user