http: move server into irc server struct to allow shutdown
Move the http server struct to be a member of the core server struct.
This allows us to shut it down when the server stops. In addition, add
an option to disable the http server, currently only used for tests. By
setting the port to null we don't launch the http server.